Major change with removal of redundant JS and HTML code.
Moved the entire implementation to a single list of dicts in the main.js
User just needs to add a route in routes.py and append the button
parameters to list in main.js. The button creation is take care via
ng-repeat. Also, settings for hiding/showing the button have been
incorporated via ng-model.

WIP:
Solve the TODOs.
Move the list from main.js to some JSON file and import.
Work on moving the settings to a modal or another webpage or popover or
bottom nav.

Signed-off-by: Diptanshu8 <diptanshuj@gmail.com>